Work

+++ src/main/kotlin/com/example/products/dto/ProductOnList.kt
@@ -3,6 +3,6 @@ package com.example.products.dto
 import com.example.models.Product
 
 @kotlinx.serialization.Serializable
-data class ProductOnList(val name: String, val price: Double) {
+data class ProductOnList(val name: String, val price: Int) {
     constructor(product: Product) : this(product.name, product.price)
 }
